Update arrow writers
