+++ src/nyx/globals.h
@@ -37,7 +37,6 @@ namespace Nyxus
 	bool processTrivialRois (const std::vector<int>& trivRoiLabels, const std::string& intens_fpath, const std::string& label_fpath, int num_FL_threads, size_t memory_limit);
 	bool processNontrivialRois (const std::vector<int>& nontrivRoiLabels, const std::string& intens_fpath, const std::string& label_fpath, int num_FL_threads);
 	void dump_roi_metrics(const std::string & label_fpath);
-	bool generate_header(ResultsCache& rescache, std::vector<std::tuple<std::string, AvailableFeatures>> F);
 
 	// Shows a message in CLI ('send_to_stderr': stdout or stderr) or Python terminal
 	void sureprint(const std::string& msg, bool send_to_stderr=false);
@@ -57,7 +56,11 @@ namespace Nyxus
 	extern const std::vector<std::string> mandatory_output_columns;
 	bool save_features_2_csv (const std::string & intFpath, const std::string & segFpath, const std::string & outputDir);
 	bool save_features_2_buffer (ResultsCache& results_cache);	
+	
 	std::vector<std::tuple<std::vector<std::string>, int, std::vector<double>>> get_feature_values();	
+	std::vector<std::string> get_header(const std::vector<std::tuple<std::string, AvailableFeatures>>& F );
+
+
 
 	void init_feature_buffers();
 	void clear_feature_buffers();
